requirements.txt
torch
torchvision
streamlit
pandas
pillow